pyproject.toml
[tool.poetry]
name = "GodLoongBot"
version = "0.1.0"
description = "Another modular bot"
authors = ["daofengql"]
license = "AGPL-3.0 license"

[tool.poetry.dependencies]
python = "^3.10"
graiax-silkcoder = {extras = ["libsndfile"], version = "^0.3.0"}
graia-saya = "^0.0.17"
graia-scheduler = "^0.1.5"
requests = "^2.28.1"
aioredis = "^2.0.1"
aiomysql = "^0.1.1"
aiocache = "^0.12.1"
aiofiles = "^0.8.0"
fastapi = "^0.79.0"
httpx = "^0.23.0"
ping3 = "^4.0.3"
numpy = "^1.23.1"
psutil = "^5.9.1"
sqlalchemy = "^1.4.39"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
aiosqlite = "^0.17.0"
graia-ariadne = "^0.11.5"
pyinstall = "^0.1.4"
goto-statement = "^1.2"
utils = "^1.0.1"
xmltojson = "^2.0.1"
azure-cognitiveservices-speech = "^1.24.2"
aio-mc-rcon = "^3.2.0"
PySSTV = "^0.5.4"
bs4 = "^0.0.1"
webdavclient3 = "^3.14.6"
pypinyin = "^0.47.1"
shodan = "^1.28.0"
lxml = "^4.9.2"
xpinyin = "^0.7.6"
b2sdk = "^1.21.0"
pydub = "^0.25.1"
soundfile = "^0.12.1"
openai = "^0.27.8"
revChatGPT = "^6.1.4"
kayaku = "^0.5.2"
Pillow = "9.3.0"
graia-amnesia = "^0.7.1"
graiax-shortcut = "^0.3.0"
redis = "^4.6.0"
minio = "^7.1.15"
mistune = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = true
